+++ tests/test_test_bigquery.py
@@ -41,3 +41,17 @@ def test_test_bigquery_complex_tables():
     print(run.pretty())
     assert run.result == "passed"
     assert all(check.result == "passed" for check in run.checks)
+
+
+@pytest.mark.skipif(
+    os.environ.get("DATACONTRACT_BIGQUERY_ACCOUNT_INFO_JSON_PATH") is None,
+    reason="Requires DATACONTRACT_BIGQUERY_ACCOUNT_INFO_JSON_PATH to be set",
+)
+def test_test_bigquery_tablename_staring_with_number():
+    data_contract = DataContract(data_contract_file="fixtures/bigquery/datacontract_tablename_starting_with_number.odcs.yaml")
+
+    run = data_contract.test()
+
+    print(run.pretty())
+    assert run.result == "passed"
+    assert all(check.result == "passed" for check in run.checks)